Move output tests from HIP_RA_Unit_tests.py to test_HIP_RA.py
